fix(server): Move `server/script/` -> `server/src/script/`
