+++ server/src/script/module/generateOpenAPIDocs.ts
