+++ server/src/script/module/modifyGeneratedAPIDocs.ts
